chore: move model to shuvi-app
